feat: favorite 컴포넌트 분리
