+++ src/components/common/artwork/ArtWork.jsx
@@ -1,9 +1,7 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import * as S from './ArtWork.style';
-import IMAGES from '../../../assets';
-import { postFavorite } from '../../../apis/postFavorite';
-import userDummy from '../../../constants/userDummy';
+import Favorite from '../../favorite/Favorite';
 
 export default function ArtWork({
   image,
@@ -14,25 +12,7 @@ export default function ArtWork({
   direction = 'row',
   size = '12rem',
 }) {
-  const [favorite, setFavorite] = useState(false);
   const url = `/product/${productId}`;
-
-  const postFav = async ({ token, userId, productId: PID }) => {
-    try {
-      const res = await postFavorite({ token, userId, PID });
-      console.log(res.result);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  const handleClick = () => {
-    console.log(favorStatus);
-    setFavorite(!favorite);
-    const userId = 24;
-    const { token } = userDummy;
-    postFav({ token, userId, productId });
-  };
 
   return (
     <S.Wrapper>
@@ -46,11 +26,7 @@ export default function ArtWork({
       </Link>
       <S.InfoWrapper>
         <S.ArtName>{title}</S.ArtName>
-        <S.Favorite
-          onClick={handleClick}
-          src={favorite ? IMAGES.favoriteOn : IMAGES.favoriteOff}
-          alt="favorite"
-        />
+        <Favorite favorStatus={favorStatus} productId={productId} />
       </S.InfoWrapper>
       <S.Artist>{authorNickname}</S.Artist>
     </S.Wrapper>
