+++ src/components/favorite/Favorite.style.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Favorite = styled.img`
+  width: 100%;
+  height: 100%;
+`;
