+++ src/constants/userDummy.js
@@ -2,7 +2,7 @@ const userDummy = {
   userId: 24,
   nickname: 'string121',
   token:
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXBlIjoiSldUIiwidXNlcl9uaWNrbmFtZSI6InN0cmluZzEyMSIsImlhdCI6MTcwNzg5NTE0MywiZXhwIjoxNzA3ODk4NzQzLCJpc3MiOiJtYWxpYnUifQ.9RSgh8qKeFJRgkWR-aKWnlAvOs1FZ-demehbcJ5GVBA',
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXBlIjoiSldUIiwidXNlcl9uaWNrbmFtZSI6InN0cmluZzEyMSIsImlhdCI6MTcwNzg5ODkwOSwiZXhwIjoxNzA3OTAyNTA5LCJpc3MiOiJtYWxpYnUifQ.Cgd7MwxqjY7kvv2KQvZmJBXYb4tGMSgOrM5uycMEZxw',
 };
 
 export default userDummy;
